+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix `Pledge` instances executing the callbacks synchronously
 * Fix `Cache#has(key)` return true for expired entries
 * Fix `memoize` Decorator not setting `max_age` when `ignore_callbacks` is true
+* Add `cast` argument to `Array#unique()`
 
 ## 0.6.3 (2018-12-06)